added mongodb capability

+++ data_processing.py
@@ -1,19 +1,21 @@
-import os
 from datetime import datetime
 
 from bs4 import BeautifulSoup
 import re
 import requests
 
-from gtfs import download_gtfs, clean_gtfs, update_version, get_version, date_format, add_to_database
+from gtfs import download_gtfs, clean_gtfs, update_version, get_version, date_format
+from database import build_database, update_data_version, get_data_version, close_database
 from utils import reset, delete_file
-from config import GTFS_FILE, DATABASE_FILE, EXTRACTED_DIRECTORY
+from config import GTFS_FILE, EXTRACTED_DIRECTORY, KEEP_TEMP_FILES
 from cloud import upload_string_to_cloud_storage
 
 # Constants
 GTFS_URL = "https://opendata.transport.vic.gov.au/dataset/gtfs-schedule"
 TRANSPORT_FILTER = "Tram"  # Can be "Metropolitan", "Tram", etc.
 KEEP_FILES = ["routes.txt", "trips.txt", "shapes.txt"]
+
+# Mongo
 
 
 def fetch_gtfs_metadata() -> tuple[datetime, str, BeautifulSoup]:
@@ -90,6 +92,8 @@ def check_if_update_needed(new_date: datetime) -> bool:
     Returns:
         bool: True if update needed, False otherwise
     """
+    print(f"Mongo old data version: {get_data_version()}")
+
     previous_date = get_version()
     print(f"Date of new data: {new_date}")
 
@@ -117,27 +121,11 @@ def download_and_extract_gtfs(download_link: str, transports_dict: dict[str,str]
     clean_gtfs(GTFS_FILE, EXTRACTED_DIRECTORY, keep_folders, KEEP_FILES)
 
 
-def build_database(transports_dict: dict[str,str]) -> None:
-    """
-    Build SQLite database from extracted GTFS files.
-
-    Args:
-        transports_dict: Dictionary of transport numbers and types
-    """
-    # Delete old database
-    delete_file(DATABASE_FILE)
-
-    # Process all extracted txt files
-    for root, dirs, files in os.walk(EXTRACTED_DIRECTORY.path):
-        for filename in files:
-            data_file_path = os.path.join(root, filename)
-
-            if data_file_path.endswith(".txt"):
-                add_to_database(DATABASE_FILE, data_file_path, transports_dict)
-
-
 def cleanup_temp_files() -> None:
     """Remove temporary files to save space."""
+    if KEEP_TEMP_FILES:
+        return
+
     delete_file(GTFS_FILE)
     delete_file(EXTRACTED_DIRECTORY)
 
@@ -159,6 +147,7 @@ def update_gtfs_data():
 
     # Update last updated date
     update_version(data_version.strftime(date_format))
+    update_data_version(data_version)
 
     # Parse transport types
     transports_dict = parse_transport_types(soup, TRANSPORT_FILTER)
@@ -188,6 +177,7 @@ def main():
     finally:
         # Test reset if needed
         reset(RESET_FILES)
+        close_database()
 
 
 if __name__ == "__main__":
